feat: documentation and CST
