+++ libs/@local/hql/ast/src/lib.rs
@@ -1,14 +1,4 @@
-pub fn add(left: u64, right: u64) -> u64 {
-    left + right
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
-}
+//! # AST for the `hql` language
+//!
+//! This is the same as the AST, it still tries to retain any code location information, but
+//! resolves any special form functions into proper AST nodes.
