+++ libs/@local/hql/cst/Cargo.toml
@@ -1,0 +1,12 @@
+cargo-features = ["edition2024"]
+
+[package]
+name = "hql-cst"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+
+[lints]
+workspace = true
